Fix bug in pair generation
This commit fixes the order of pairs generated. Previously, the strong
suit would be kept on the left.

This rework also changes variable names for easier pattern recognition
when reading the code. strong ranks and suits are now generated on the
earliest iterations of the loops to more closely match their
`enumerate` starts.